requirements.txt
bidict==0.22.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
eventlet==0.33.3
filelock==3.10.7
Flask==2.2.3
Flask-SocketIO==5.3.3
greenlet==2.0.2
huggingface-hub==0.13.3
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
numpy==1.24.2
packaging==23.0
python-engineio==4.4.0
python-socketio==5.8.0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
six==1.16.0
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.27.4
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wincertstore==0.2
zipp==3.15.0
